Revert "Cb in validate"
