+++ components/NewEventForm.js
@@ -27,60 +27,6 @@ function numToTime(num) {
   return str;
 }
 
-// converts time hh:mm AM/PM to minute of the day
-function convertTime(str) {
-  var min = 0;
-  var morn = 0; // 0 if morning, 1 if afternoon
-  if (str.charAt(str.length - 2) == "P") {
-    morn = 1;
-  }
-  str = str.slice(0, -3);
-  var str_split = str.split(":");
-
-  if (str_split[0] == 12) {
-    str_split[0] = 0;
-  }
-
-  min = +str_split[0] * 60 + +str_split[1];
-
-  if (morn == 1) {
-    min = min + 720;
-  }
-  return min;
-}
-
-function validateForm(
-  e,
-  isSunday,
-  isMonday,
-  isTuesday,
-  isWednesday,
-  isThursday,
-  isFriday,
-  isSaturday,
-  startTime,
-  endTime,
-  addEvent
-) {
-  if (convertTime(startTime) >= convertTime(endTime)) {
-    window.alert("Start time must be before end time.");
-    return false;
-  }
-  if (
-    !isSunday &&
-    !isMonday &&
-    !isTuesday &&
-    !isWednesday &&
-    !isThursday &&
-    !isFriday &&
-    !isSaturday
-  ) {
-    window.alert("Must select at least one day of the week.");
-    return false;
-  }
-  return addEvent(e);
-}
-
 function NewEventForm() {
   const { showToast } = useToasts();
   const { mutate } = useSWR("/api/event");
@@ -181,24 +127,7 @@ function NewEventForm() {
         </Accordion.Toggle>
         <Accordion.Collapse eventKey="0">
           <Card.Body>
-            <Form
-              onSubmit={(e) =>
-                validateForm(
-                  e,
-                  isSunday,
-                  isMonday,
-                  isTuesday,
-                  isWednesday,
-                  isThursday,
-                  isFriday,
-                  isSaturday,
-                  startTime,
-                  endTime,
-                  addEvent
-                )
-              }
-              className="mb-3"
-            >
+            <Form onSubmit={addEvent} className="mb-3">
               <Form.Group>
                 <Container>
                   <Row>
