+++ pages/my-group/[code]/index.js
@@ -1,9 +1,10 @@
 import Layout from "../../../components/Layout";
-import JoinLeaveButton from "../../../components/JoinLeaveButton";
 import { requiredAuth } from "../../../utils/ssr";
 import { useRouter } from "next/router";
 import useSWR from "swr";
-import MembersList from "../../../components/MembersList";
+import JoinButton from "./JoinButton";
+import LeaveButton from "./LeaveButton";
+import MembersList from "./MembersList";
 import Button from "react-bootstrap/Button";
 import GroupFreeTime from "../../../components/GroupFreeTime";
 import MembersFreeTime from "../../../components/MembersFreeTime";
@@ -31,7 +32,8 @@ export function GroupPage(props) {
           <h2>Group Code: {code}</h2>
           <Form inline>
             <Form.Group>
-              <JoinLeaveButton />
+              <JoinButton />
+              <LeaveButton />
             </Form.Group>
           </Form>
           <MembersList />
