--- stories/MembersFreeTime.stories.js
@@ -1,11 +1,0 @@
-import React from "react";
-import MembersFreeTime from "../components/MembersFreeTime";
-
-export default {
-  title: "MembersFreeTime",
-  component: MembersFreeTime,
-};
-
-export const noMembers = () => {
-  return <MembersFreeTime />;
-};
