--- stories/MembersList.stories.js
@@ -1,12 +1,0 @@
-import React from "react";
-import { text } from "@storybook/addon-knobs";
-import MembersList from "../components/MembersList";
-
-export default {
-  title: "MembersList",
-  component: MembersList,
-};
-
-export const noMembers = () => {
-  return <MembersList />;
-};
